backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
openai
python-dotenv